django_test/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
drf-nested-routers==0.93.4
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2